requirements.txt
python-binance==1.0.10
colorama==0.4.4
PyYAML==5.4.1
verification-code==1.0.0
mysql-connector-python==8.0.25